+++ Assets/3.Script/Player/PlayerBullet.cs
@@ -9,7 +9,7 @@ using UnityEngine;
 /// </summary>
 public class PlayerBullet : MonoBehaviour, ICollisionDamage
 {
-    public int CollisionDamage { get; set; }
+    public float CollisionDamage { get; set; }
 
     public void OnTriggerEnter2D(Collider2D collision)
     {
